AVR-01
"master" of the Alexa/Amazon side of the application.
Implemented a first Skill version with basic responses and persistence with S3.
Cloudwatch is enabled as well.
All is hosted in the eu-west-1, this must remain because Alexa self choose the region by language (this one is for italian).
A basic websocket interface is activated as well
All of this is made possible to work with thanks to serverless (node package).
Updated gitignore for macOS and nodejs